rename:(cookbook-state) use 'cookbooks' cmd
